requirements.txt
discord.py>=2.3.2
openai>=1.40.0
python-dotenv>=1.0.1
requests>=2.32.0
